fix: code review - remove btc client from tx builder - move utxo selection by value to utxo manager - remove un used imports
